requirements.txt
scraperwiki
lxml
cssselect
requests-html
pandas